[ENH] [REFACTOR] imports to improve performance of the cli.
