+++ nipoppy_cli/docs/source/conf.py
@@ -95,7 +95,7 @@ nitpick_ignore = [
     ("py:class", "argparse._SubParsersAction"),
     ("py:class", "argparse._ActionsContainer"),
     ("py:class", "StrOrPathLike"),
-    ("py:class", "nipoppy.utils.StrOrPathLike"),
+    ("py:class", "nipoppy.env.StrOrPathLike"),
     ("py:class", "typing_extensions.Self"),
 ]
