+++ nipoppy_cli/nipoppy/config/container.py
@@ -11,7 +11,7 @@ from typing import Any, Optional
 from pydantic import BaseModel, ConfigDict, Field
 
 from nipoppy.logger import get_logger
-from nipoppy.utils import StrOrPathLike
+from nipoppy.env import StrOrPathLike
 
 # Apptainer
 APPTAINER_BIND_FLAG = "--bind"
