+++ nipoppy_cli/nipoppy/workflows/doughnut.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Optional
 
 from nipoppy.tabular.doughnut import Doughnut, generate_doughnut, update_doughnut
-from nipoppy.utils import StrOrPathLike
+from nipoppy.env import StrOrPathLike
 from nipoppy.workflows.base import BaseWorkflow
