+++ nipoppy_cli/nipoppy/workflows/tracker.py
@@ -7,7 +7,8 @@ from pydantic import TypeAdapter
 
 from nipoppy.config.tracker import TrackerConfig, check_tracker_configs
 from nipoppy.tabular.bagel import Bagel
-from nipoppy.utils import StrOrPathLike, load_json
+from nipoppy.utils import load_json
+from nipoppy.env import StrOrPathLike
 from nipoppy.workflows.pipeline import BasePipelineWorkflow
