+++ nipoppy_cli/tests/test_tabular_doughnut.py
@@ -7,7 +7,7 @@ import pytest
 
 from nipoppy.tabular.dicom_dir_map import DicomDirMap
 from nipoppy.tabular.doughnut import Doughnut, generate_doughnut, update_doughnut
-from nipoppy.utils import StrOrPathLike
+from nipoppy.env import StrOrPathLike
 
 from .conftest import DPATH_TEST_DATA, check_doughnut, prepare_dataset
